// services/schemas/inputTypeSchemas/ProductUpdateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { FloatFieldUpdateOperationsInputSchema } from './FloatFieldUpdateOperationsInputSchema';
import { IntFieldUpdateOperationsInputSchema } from './IntFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { UserUpdateOneRequiredWithoutProductNestedInputSchema } from './UserUpdateOneRequiredWithoutProductNestedInputSchema';
import { QuantityUpdateManyWithoutProductNestedInputSchema } from './QuantityUpdateManyWithoutProductNestedInputSchema';
import { CategoryUpdateOneWithoutProductsNestedInputSchema } from './CategoryUpdateOneWithoutProductsNestedInputSchema';

export const ProductUpdateInputSchema: z.ZodType<Prisma.ProductUpdateInput> = z.object({
  id: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  description: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  image: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  price: z.union([ z.number(),z.lazy(() => FloatFieldUpdateOperationsInputSchema) ]).optional(),
  stock: z.union([ z.number().int(),z.lazy(() => IntFieldUpdateOperationsInputSchema) ]).optional(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  Vendor: z.lazy(() => UserUpdateOneRequiredWithoutProductNestedInputSchema).optional(),
  Quantity: z.lazy(() => QuantityUpdateManyWithoutProductNestedInputSchema).optional(),
  Category: z.lazy(() => CategoryUpdateOneWithoutProductsNestedInputSchema).optional()
}).strict();

export default ProductUpdateInputSchema;